// packages/core/src/types.d.ts
export type GridRowId = string | number;
export type GridRowModelType = "client" | "server";
export type SortDirection = "asc" | "desc";
export type ColumnPin = "left" | "right";
export type FilterOperator = "contains" | "equals" | "startsWith" | "endsWith" | "gt" | "gte" | "lt" | "lte" | "between" | "isEmpty" | "isNotEmpty" | "in";
export type Accessor<TRow, TValue = unknown> = (row: TRow) => TValue;
export type ColumnComparator<TValue = unknown, TRow = unknown> = (leftValue: TValue, rightValue: TValue, leftRow: TRow, rightRow: TRow) => number;
export type ColumnFilterPredicate<TValue = unknown, TRow = unknown> = (value: TValue, filter: FilterRule, row: TRow) => boolean;
export type ValueFormatter<TValue = unknown, TRow = unknown> = (value: TValue, row: TRow) => string;
export type ValueParser<TValue = unknown, TRow = unknown> = (input: string, row: TRow) => TValue;
export type ColumnDef<TRow, TValue = unknown> = {
    id?: string;
    field?: Extract<keyof TRow, string> | string;
    headerName?: string;
    headerGroup?: string;
    accessor?: Accessor<TRow, TValue>;
    sortable?: boolean;
    filterable?: boolean;
    editable?: boolean;
    hidden?: boolean;
    pinned?: ColumnPin;
    width?: number;
    minWidth?: number;
    maxWidth?: number;
    comparator?: ColumnComparator<TValue, TRow>;
    filterPredicate?: ColumnFilterPredicate<TValue, TRow>;
    valueFormatter?: ValueFormatter<TValue, TRow>;
    valueParser?: ValueParser<TValue, TRow>;
};
export type ResolvedColumnDef<TRow, TValue = unknown> = Omit<ColumnDef<TRow, TValue>, "id" | "accessor"> & {
    id: string;
    headerName: string;
    accessor: Accessor<TRow, TValue>;
};
export type SortRule = {
    columnId: string;
    direction: SortDirection;
};
export type ColumnState = {
    columnId: string;
    hidden?: boolean;
    pinned?: ColumnPin;
    width?: number;
    order?: number;
};
export type FilterRule = {
    columnId: string;
    operator: FilterOperator;
    value?: unknown;
};
// ...
export type RowNode<TRow> = {
    id: GridRowId;
    index: number;
    original: TRow;
};